implement apply_fn_with_args, split string manipulation functions into a new namespace
